Fixed some bugs.

+++ src/main/kotlin/cs309/dormiselect/backend/data/teacher/TeamMemberRemoveDto.kt
@@ -1,3 +1,3 @@
 package cs309.dormiselect.backend.data.teacher
 
-data class TeamMemberRemoveDto(val teamId: Int, val studentId: Long)
+data class TeamMemberRemoveDto(val teamId: Int, val studentId: Int)
