+++ src/main/kotlin/cs309/dormiselect/backend/repo/StudentRepo.kt
@@ -2,7 +2,6 @@ package cs309.dormiselect.backend.repo
 
 import cs309.dormiselect.backend.domain.account.Student
 import org.springframework.data.domain.Page
-import org.springframework.data.domain.PageRequest
 import org.springframework.data.domain.Pageable
 import org.springframework.data.repository.CrudRepository
 
@@ -12,8 +11,8 @@ interface StudentRepo : CrudRepository<Student, Int> {
 
     fun findAll(pageable: Pageable): Page<Student>
 
-    fun existsByStudentId(studentId: Long): Boolean
-    fun findByStudentId(studentId: Long):Student?
+    fun existsByStudentId(studentId: Int): Boolean
+    fun findByStudentId(studentId: Int): Student?
 
 
 }
